move location.java
Update AmateurSondehubTest.java
