+++ ttnhabbridge/src/main/java/nl/sikken/bertrik/TtnHabBridge.java
@@ -19,11 +19,11 @@ import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import nl.sikken.bertrik.hab.DecodeException;
 import nl.sikken.bertrik.hab.EPayloadEncoding;
 import nl.sikken.bertrik.hab.ExpiringCache;
+import nl.sikken.bertrik.hab.Location;
 import nl.sikken.bertrik.hab.PayloadDecoder;
 import nl.sikken.bertrik.hab.Sentence;
 import nl.sikken.bertrik.hab.habitat.HabReceiver;
 import nl.sikken.bertrik.hab.habitat.HabitatUploader;
-import nl.sikken.bertrik.hab.habitat.Location;
 import nl.sikken.bertrik.hab.lorawan.HeliumUplinkMessage;
 import nl.sikken.bertrik.hab.lorawan.LoraWanUplinkMessage;
 import nl.sikken.bertrik.hab.lorawan.LoraWanUplinkMessage.GatewayInfo;
