+++ ttnhabbridge/src/main/java/nl/sikken/bertrik/hab/Location.java
@@ -1,4 +1,4 @@
-package nl.sikken.bertrik.hab.habitat;
+package nl.sikken.bertrik.hab;
 
 /**
  * Representation of a HAB receiver location.
