+++ ttnhabbridge/src/test/java/nl/sikken/bertrik/hab/habitat/HabitatUploaderTest.java
@@ -9,6 +9,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 import org.mockito.Mockito;
 
+import nl.sikken.bertrik.hab.Location;
 import nl.sikken.bertrik.hab.Sentence;
 import retrofit2.mock.Calls;
